+++ src/libaktualizr/package_manager/dockerappmanager_test.cc
@@ -82,7 +82,7 @@ TEST(DockerAppManager, DockerAppStandalone) {
   std::shared_ptr<INvStorage> storage = INvStorage::newStorage(config.storage);
   KeyManager keys(storage, config.keymanagerConfig());
   auto client = std_::make_unique<SotaUptaneClient>(config, storage);
-  ASSERT_TRUE(client->updateImagesMeta());
+  ASSERT_TRUE(client->updateImageMeta());
 
   std::string targets = Utils::readFile(repo / "repo/repo/targets.json");
   LOG_INFO << "Repo targets " << targets;